LIDE-002: Added Vagrant synch dir
Added the LIDE repo as a synched directory in Vagrant

PROS: We'll need access to the repo inside the CDE and this is a very
easy way to do it.  Plus, it assists with the transfer of files between
the CDE and the host.

CONS: I know from past experiences that there are ghosts in the machine,
tied to Vagrant's synch dir, when the host OS is Windows and the guest
OS is Linux.  B E   W A R Y

Changes to be committed:
	modified:   Vagrantfile