LIDE-002: Added Gnome GTK tasks
This took some doing to get Ansible to change the desktop theme to dark.
The trick?  I had to install the theme first?!  Whatever.

I tested this with a vagrant destroy just to ensure I didn't *manually*
set this in my CDE.  It appears to work.

Changes to be committed:
	modified:   devops/ansible/provision-cde.yml
	new file:   devops/ansible/tasks/linux/tweak-gnome.yml