LIDE-004: Refactored challenge 101 solution
Refactored the challenge 101 example solution so that the example signal
handler was "reentrant" and "async-signal-safe".

Also, swapped out all calls to printf-family functions to write() calls
so the test class can check for non-reentrant and/or non-async-signal-safe
function calls (like it does for calls to signal()).

Manual testing passes.
ASAN is happy.
Let's see how it holds up against the test cases.

Changes to be committed:
	modified:   solution/101_example.c